+++ tests/testthat/test_print.summary.nprcmanagGV.R
@@ -1,10 +1,13 @@
 context("print.summary.nprcmanagGV")
 library(testthat)
-library(lubridate)
-library(stringi)
 
-set_seed(10)
 test_that("print.summary.nprcmanagGV prints expected output", {
-  pedClean <- reportGV(nprcmanager::qcPed)
-  expect_success(expect_output(print(summary(pedClean))))
-  })
+  ped <- nprcmanager::pedOne
+  ped$birth_date[ped$ego_id == "d2"] <- "2000-04-13"
+  ped$birth_date[ped$ego_id == "o4"] <- "2016-04-13"
+  ped <- suppressWarnings(qcStudbook(ped, reportErrors = FALSE))
+  summaryGV <- summary(reportGV(ped, guIter = 10))
+  expect_success(expect_output(print(summaryGV)))
+  expect_equal(summaryGV[1], c("The genetic value report"))
+  expect_equal(length(summaryGV), 8)
+})