+++ NEWS.md
@@ -3,6 +3,11 @@ NEWS
 R. Mark Sharp
 06/24/2019
 
+# nprcmanager 0.5.22 (20190624)
+
+  - Corrected and augmented unit tests for print\_summary\_nprcmanagGV
+    and summary.nprcmanagGV
+
 # nprcmanager 0.5.21 (20190624)
 
   - Added unit tests for print\_summary\_nprcmanagGV and